Calculate validation size

From and expected p and desired std_err, it returns the minimal validation size for your assesment sets or validation sets.

Usage

calc_validation_size(expected_p, desired_std_err)

Source

https://stats.stackexchange.com/a/304996/7387

Arguments

expected_p

An object

desired_std_err

An expresion

Details

The calc_validation_size function returns the minimal validation size for expected probabilities and desired error.

See also

Examples

calc_validation_size(expected_p = 0.8, desired_std_err = 0.02)
#> [1] 400