Do the last fit and get the metrics

Do the last fit and get the metrics

last_metrics(res, split, metric)

Arguments

res

Tune results

split

The initial split object

metric

What metric to use to select the best workflow

See also

Examples

library(tidymodels)
#> ── Attaching packages ────────────────────────────────────── tidymodels 1.2.0 ──
#> ✔ broom        1.0.6           ✔ recipes      1.0.10.9000
#> ✔ dials        1.2.1           ✔ rsample      1.2.1      
#> ✔ dplyr        1.1.4.9000      ✔ tibble       3.2.1      
#> ✔ ggplot2      3.5.1           ✔ tidyr        1.3.1      
#> ✔ infer        1.0.7           ✔ tune         1.2.1.9000 
#> ✔ modeldata    1.3.0           ✔ workflows    1.1.4.9000 
#> ✔ parsnip      1.2.1.9001      ✔ workflowsets 1.1.0      
#> ✔ purrr        1.0.2           ✔ yardstick    1.3.1      
#> ── Conflicts ───────────────────────────────────────── tidymodels_conflicts() ──
#> ✖ purrr::discard() masks scales::discard()
#> ✖ dplyr::filter()  masks stats::filter()
#> ✖ dplyr::lag()     masks stats::lag()
#> ✖ recipes::step()  masks stats::step()
#> • Dig deeper into tidy modeling with R at https://www.tmwr.org
library(xgboost)
#> 
#> Attaching package: ‘xgboost’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     slice
library(modeldata)
data(cells)
split <- cells |>
  mutate(across(where(is.character), as.factor)) |>
  sample_n(500) |>
  initial_split(strata = class)
train <- training(split)
folds <- vfold_cv(train, v = 2, strata = class)
wf <- train |>
  recipe(case ~ .) |>
  step_integer(all_nominal_predictors())  |>
  workflow_boost_tree()
res <- wf |>
  tune_grid(
    folds,
    grid = 2,
    metrics = metric_set(roc_auc),
    control = control_grid(save_workflow = TRUE, verbose = FALSE)
  )
#> i Creating pre-processing data to finalize unknown parameter: mtry
res |> collect_metrics()
#> # A tibble: 2 × 13
#>    mtry trees min_n tree_depth learn_rate loss_reduction sample_size .metric
#>   <int> <int> <int>      <int>      <dbl>          <dbl>       <dbl> <chr>  
#> 1    14   849    23         14    0.238      0.0172            0.189 roc_auc
#> 2    37  1324     9          5    0.00199    0.000000173       0.554 roc_auc
#> # ℹ 5 more variables: .estimator <chr>, mean <dbl>, n <int>, std_err <dbl>,
#> #   .config <chr>
res |> last_metrics(split, "roc_auc")
#> # A tibble: 3 × 4
#>   .metric     .estimator .estimate .config             
#>   <chr>       <chr>          <dbl> <chr>               
#> 1 accuracy    binary         0.568 Preprocessor1_Model1
#> 2 roc_auc     binary         0.5   Preprocessor1_Model1
#> 3 brier_class binary         0.247 Preprocessor1_Model1
best <- res |> fit_best()
best |>
  augment(testing(split)) |>
  roc_auc(case, .pred_Test) |>
  pull(.estimate)
#> [1] 0.5