Prep, juice and glimpse a recipe or workflow

Prep, juice and glimpse a recipe or workflow

prep_juice(object)

Source

https://recipes.tidymodels.org/reference/update.step.html

Arguments

object

A recipe or a workflow object with a recipe

See also

Examples

recipes::recipe(spray ~ ., data = InsectSprays) |>
   prep_juice()
#> Rows: 72
#> Columns: 2
#> $ count <dbl> 10, 7, 20, 14, 14, 12, 10, 23, 17, 20, 14, 13, 11, 17, 21, 11, 1…
#> $ spray <fct> A, A, A, A, A, A, A, A, A, A, A, A, B, B, B, B, B, B, B, B, B, B…
recipes::recipe(spray ~ ., data = InsectSprays)  |>
   workflows::workflow(parsnip::linear_reg()) |>
   prep_juice()
#> Rows: 72
#> Columns: 2
#> $ count <dbl> 10, 7, 20, 14, 14, 12, 10, 23, 17, 20, 14, 13, 11, 17, 21, 11, 1…
#> $ spray <fct> A, A, A, A, A, A, A, A, A, A, A, A, B, B, B, B, B, B, B, B, B, B…