Create an xgboost tunable workflow for regression and classification

Create an xgboost tunable workflow for regression and classification

workflow_boost_tree(rec, engine = "xgboost", counts = TRUE, ...)

Arguments

rec

prerocessing recipe to build the workflow

engine

optional, xgboost by default

counts

Optional logic argument wether mtry use counts or not

...

optional engine arguments

See also

Examples

library(tidymodels)
library(xgboost)
library(modeldata)
library(future)
data(cells)
split <- cells |>
  mutate(across(where(is.character), as.factor)) |>
  sample_n(500) |>
  initial_split(strata = class)
train <- training(split)
folds <- vfold_cv(train, v = 2, strata = class)
wf <- train |>
  recipe(case ~ .) |>
  step_integer(all_nominal_predictors())  |>
  workflow_boost_tree()
doFuture::registerDoFuture()
plan(sequential)
res <- wf |>
  tune_grid(
    folds,
    grid = 2,
    metrics = metric_set(roc_auc),
    control = control_grid(save_workflow = TRUE, verbose = FALSE)
  )
#> i Creating pre-processing data to finalize unknown parameter: mtry
res |> collect_metrics()
#> # A tibble: 2 × 13
#>    mtry trees min_n tree_depth learn_rate loss_reduction sample_size .metric
#>   <int> <int> <int>      <int>      <dbl>          <dbl>       <dbl> <chr>  
#> 1     3  1855     9          1    0.127       0.00000368       0.264 roc_auc
#> 2    37   920    32         12    0.00136     2.13             0.951 roc_auc
#> # ℹ 5 more variables: .estimator <chr>, mean <dbl>, n <int>, std_err <dbl>,
#> #   .config <chr>
res |> last_metrics(split, "roc_auc")
#> # A tibble: 3 × 4
#>   .metric     .estimator .estimate .config             
#>   <chr>       <chr>          <dbl> <chr>               
#> 1 accuracy    binary         0.444 Preprocessor1_Model1
#> 2 roc_auc     binary         0.446 Preprocessor1_Model1
#> 3 brier_class binary         0.252 Preprocessor1_Model1
best <- res |> fit_best()
best |>
  augment(testing(split)) |>
  roc_auc(case, .pred_Test) |>
  pull(.estimate)
#> [1] 0.4461946